Refactor: import 하는 Script 변경에 따른 Script 사용 파일 수정
