+++ src/main/java/com/echall/platform/content/domain/entity/ContentEntity.java
@@ -4,6 +4,7 @@ import com.echall.platform.category.domain.entity.CategoryEntity;
 import com.echall.platform.content.domain.dto.ContentRequestDto;
 import com.echall.platform.content.domain.enums.ContentStatus;
 import com.echall.platform.content.domain.enums.ContentType;
+import com.echall.platform.script.domain.entity.Script;
 import com.echall.platform.user.domain.entity.BaseEntity;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.Size;
@@ -64,7 +65,7 @@ public class ContentEntity extends BaseEntity {
 		this.contentType = contentType;
 		this.mongoContentId = mongoContentId;
 		this.preScripts = truncate(
-			preScripts.subList(0, 5)
+			preScripts.subList(0, Math.min(preScripts.size(), 5))
 				.stream()
 				.map(Script::getEnScript)
 				.toList().toString()
@@ -78,7 +79,7 @@ public class ContentEntity extends BaseEntity {
 		this.title = dto.title();
 		this.contentStatus = dto.contentStatus() == null ? ContentStatus.ACTIVATED : dto.contentStatus();
 		this.preScripts = truncate(
-			dto.script().subList(0, 5)
+			dto.script().subList(0, Math.min(dto.script().size(), 5))
 				.stream()
 				.map(Script::getEnScript)
 				.toList().toString()
